Create Challenge1_Python
Practice challenge #1 from http://www.practicepython.org/
Create a program that asks the user to enter their name and their age.  Print out a message addressed to them that tells them the year that they will turn 100 years old.
Extras: Add on to the previous program by asking the user for another number and printing out that many copies of the previous message.  Print out that many copies of the previous message on separate lines.